requirements.txt
Django==5.2.6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient        # Windows 下若安装困难，可改用 PyMySQL（见下方说明）
